+++ server/src/middleware/websocket.adapter.ts
@@ -3,7 +3,7 @@ import { IoAdapter } from '@nestjs/platform-socket.io';
 import { createAdapter } from '@socket.io/redis-adapter';
 import { Redis } from 'ioredis';
 import { ServerOptions } from 'socket.io';
-import { IConfigRepository } from 'src/interfaces/config.interface';
+import { ConfigRepository } from 'src/repositories/config.repository';
 
 export class WebSocketAdapter extends IoAdapter {
   constructor(private app: INestApplicationContext) {
@@ -11,7 +11,7 @@ export class WebSocketAdapter extends IoAdapter {
   }
 
   createIOServer(port: number, options?: ServerOptions): any {
-    const { redis } = this.app.get<IConfigRepository>(IConfigRepository).getEnv();
+    const { redis } = this.app.get(ConfigRepository).getEnv();
     const server = super.createIOServer(port, options);
     const pubClient = new Redis(redis);
     const subClient = pubClient.duplicate();
