+++ server/src/repositories/map.repository.ts
@@ -11,7 +11,6 @@ import { DB, GeodataPlaces, NaturalearthCountries } from 'src/db';
 import { AssetEntity, withExif } from 'src/entities/asset.entity';
 import { NaturalEarthCountriesTempEntity } from 'src/entities/natural-earth-countries.entity';
 import { LogLevel, SystemMetadataKey } from 'src/enum';
-import { IConfigRepository } from 'src/interfaces/config.interface';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import {
   GeoPoint,
@@ -21,6 +20,7 @@ import {
   ReverseGeocodeResult,
 } from 'src/interfaces/map.interface';
 import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
+import { ConfigRepository } from 'src/repositories/config.repository';
 
 interface MapDB extends DB {
   geodata_places_tmp: GeodataPlaces;
@@ -30,7 +30,7 @@ interface MapDB extends DB {
 @Injectable()
 export class MapRepository implements IMapRepository {
   constructor(
-    @Inject(IConfigRepository) private configRepository: IConfigRepository,
+    private configRepository: ConfigRepository,
     @Inject(ISystemMetadataRepository) private metadataRepository: ISystemMetadataRepository,
     @Inject(ILoggerRepository) private logger: ILoggerRepository,
     @InjectKysely() private db: Kysely<MapDB>,
