+++ server/src/services/backup.service.spec.ts
@@ -2,7 +2,6 @@ import { PassThrough } from 'node:stream';
 import { defaults, SystemConfig } from 'src/config';
 import { StorageCore } from 'src/cores/storage.core';
 import { ImmichWorker, StorageFolder } from 'src/enum';
-import { IConfigRepository } from 'src/interfaces/config.interface';
 import { ICronRepository } from 'src/interfaces/cron.interface';
 import { IDatabaseRepository } from 'src/interfaces/database.interface';
 import { JobStatus } from 'src/interfaces/job.interface';
@@ -10,6 +9,7 @@ import { IProcessRepository } from 'src/interfaces/process.interface';
 import { IStorageRepository } from 'src/interfaces/storage.interface';
 import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
 import { BackupService } from 'src/services/backup.service';
+import { IConfigRepository } from 'src/types';
 import { systemConfigStub } from 'test/fixtures/system-config.stub';
 import { mockSpawn, newTestService } from 'test/utils';
 import { describe, Mocked } from 'vitest';
