+++ server/src/services/job.service.spec.ts
@@ -2,11 +2,11 @@ import { BadRequestException } from '@nestjs/common';
 import { defaults, SystemConfig } from 'src/config';
 import { ImmichWorker } from 'src/enum';
 import { IAssetRepository } from 'src/interfaces/asset.interface';
-import { IConfigRepository } from 'src/interfaces/config.interface';
 import { IJobRepository, JobCommand, JobItem, JobName, JobStatus, QueueName } from 'src/interfaces/job.interface';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { ITelemetryRepository } from 'src/interfaces/telemetry.interface';
 import { JobService } from 'src/services/job.service';
+import { IConfigRepository } from 'src/types';
 import { assetStub } from 'test/fixtures/asset.stub';
 import { newTestService } from 'test/utils';
 import { Mocked } from 'vitest';
