+++ server/src/utils/config.ts
@@ -6,10 +6,10 @@ import * as _ from 'lodash';
 import { SystemConfig, defaults } from 'src/config';
 import { SystemConfigDto } from 'src/dtos/system-config.dto';
 import { SystemMetadataKey } from 'src/enum';
-import { IConfigRepository } from 'src/interfaces/config.interface';
 import { DatabaseLock } from 'src/interfaces/database.interface';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
+import { IConfigRepository } from 'src/types';
 import { getKeysDeep, unsetDeep } from 'src/utils/misc';
 import { DeepPartial } from 'typeorm';
