+++ server/test/repositories/config.repository.mock.ts
@@ -1,8 +1,9 @@
 import { PostgresJSDialect } from 'kysely-postgres-js';
 import postgres from 'postgres';
 import { ImmichEnvironment, ImmichWorker } from 'src/enum';
-import { EnvData, IConfigRepository } from 'src/interfaces/config.interface';
 import { DatabaseExtension } from 'src/interfaces/database.interface';
+import { EnvData } from 'src/repositories/config.repository';
+import { IConfigRepository } from 'src/types';
 import { Mocked, vitest } from 'vitest';
 
 const envData: EnvData = {
